move scripts relayer
