+++ relayer/tsconfig.json
@@ -15,7 +15,7 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true
   },
-  "include": ["src", "closeUpdateState.ts", "nukeDB.ts"],
+  "include": ["src"],
   "dry": false,
   "debug": false
 }
